+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/LightPunctual.cs
@@ -4,8 +4,6 @@
 using System;
 using Unity.Gltfast.Text.Json.Serialization;
 
-using UnityEngine.Assertions;
-
 namespace GLTFast.Schema
 {
     /// <summary>