+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/Material.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using Unity.Gltfast.Text.Json;
 using Unity.Gltfast.Text.Json.Serialization;
-using Unity.Mathematics;
 using UnityEngine;
 
 namespace GLTFast.Schema