+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/MaterialSpecular.cs
@@ -3,7 +3,6 @@
 
 using System;
 using Unity.Gltfast.Text.Json.Serialization;
-using Unity.Mathematics;
 using UnityEngine;
 
 namespace GLTFast.Schema