+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/OcclusionTextureInfo.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using Unity.Gltfast.Text.Json.Serialization;
-using Unity.Mathematics;
 
 namespace GLTFast.Schema
 {