+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/PbrMetallicRoughness.cs
@@ -4,9 +4,7 @@
 using System.Collections.Generic;
 using Unity.Gltfast.Text.Json;
 using Unity.Gltfast.Text.Json.Serialization;
-using Unity.Mathematics;
 using UnityEngine;
-using UnityEngine.Serialization;
 
 namespace GLTFast.Schema
 {